requirements.txt


Django<4
djangorestframework~=3.13.0
pydantic<=1.9.2